fix: Visitor class NPE
